Improved implementation of image insertion into images array
Before this fix, images were loading from the front of the array to the back, resulting in the user seeing a flood of images loading consecutively upon load.

Now the insertion looks for the first non-nil value in the array, inserts the new image there (rather than at index 0), resulting in the viewer seeing the images load consecutively.